fix: fault injection terminates worker threads
We were calling `sys.exit(1)` to terminate a worker thread and
(hopefully) generate a `connection reset by peer` (or similar) error on
the client. This is not needed, as we explicitly close the socket a few
lines before. And it is harmful, because it terminates the worker
threads and eventually exhausts them.

With this change we just close the socket, and introduce a new test to
verify one can use the `return-broken-stream` or
`return-reset-connection` instructions multiple times.