cleanup: simplify pre-condition handling
This simplifies pre-condition handling for `Database.get_object()` and
`Database.delete_object()` by asking the caller to provide the
generation (which can be `None` or `0`) and the preconditions as a list
of predicates on the blob and the live generation.

The callers know what the pre-conditions fields are called (e.g.
`ifGenerationMatch` vs. `ifSourceGenerationMatch` for `Objects: copy` or
`Objects: rewrite`). The caller also knows if the generation is a
top-level attribute of the request, or part of a sub-object (or a query
parameter or some other possibilities).