cleanup: simplify object precondition handling
This simplifies precondition handling for `Database.*_object()`. The
caller needs to provide the generation (which can be `None` or `0`) and
the preconditions as a list of predicates on the blob and the live
generation.

In general, this makes the code more readable, albeit more verbose,
because the caller knows how the preconditions are represented in the
different transports (e.g. headers in XML, query parameters in JSON,
proto fields in gRPC), and in some cases the preconditions have
different names (e.g. `ifSource*` for `Objects: copy` or `Objects:
rewrite`).