fix: object deletes with a generation
I accidentally broke object deletes when a generation number is
provided.  We had no tests for this (now we do). Turns out in REST the
generation was a string, and the code was expecting an `int`. I added a
test (that reproduced the problem), and also some type annotations
(which seem to help very little).