fix(grpc): more problems with UpdateObject()/UpdateBucket()
I thought my previous tests had reproduced and fixed the problem, but
evidently this is not the case.  It all seems to be working correctly
now, tested using the C++ client library.  I also filed a bug against
the protobuf project, and another bug to cleanup this code when the
protobuf bug is fixed.