fix: Correctly handle aborts with multiple ranges. (#16)
We never polled the `responses` queue after an `"abort"` message was
dequeued. That can stop us from ever joining `gather_thread` if another
range enqueues a response first, because the queue capacity is fixed. We
want to keep the queue capacity fixed so that we can increase the
likelihood of interleaving responses from multiple ranges.

Fix by draining the queue in the `finally` block before reraising.

Also, correct `abort` to `abort_with_status`: if the error is a `grpc.Status`,
we're meant to call `context.abort_with_status` rather than
`context.abort`.