feat: Handle appendable objects in BidiWriteObject (#18)
Appendable objects extend BidiWriteObject calls. This is a refactor and
a mostly-accurate implementation of appendable objects in the testbench.

There's a key difference - this simulates true appendable using
resumable uploads, which are not visible in ListObjects or similar calls
in the same way that appendable objects are. That's sufficient for
current testing.