chore: Timeout with cancellation in BidiReadObject (#42)
We want to limit the total amount of time any one request can monopolize
a thread on the gRPC server, so we have a 10 second timer to terminate
the request. However, the existing code ends the stream with OK, which
is misleading to clients.

By switching to cancellation, we no longer need to set a timeout on the
gather thread join (since it observes an error from
next(request_iterator)) and we return a CANCELLED error which correctly
indicates to client code that they probably wrote a bug.

Now, tests with bugs that keep the stream open for more than 10 seconds
look like:

```
=== RUN   TestRetryConformance/grpc-1-[return-reset-connection_return-reset-connection]-storage.objects.get-1
    retry_conformance_test.go:696: want success, got rpc error: code = Canceled desc = CANCELLED
```

Tests pass when code correctly cleans up channels/RPCs.

N.B. if we ever need longer than 10 second RPCs and we _don't_ want them
to end in a CANCELLED error, we will have to revisit this.