fix: Disable pipelining in BidiReadObject. (#43)
The real implementation of BidiReadObject may pipeline requests,
concurrently serving reads issued in separate messages. Unfortunately,
as best we can tell, gRPC in synchronous Python makes it hard to abort
the next(request_iterator) call when there's an error with a concurrent
read.

So we go back to a previous iteration of this code, which handled
messages in batches. We no longer need a response queue or a thread
pool.